namedspaced mode setting
