+++ lib/instructor.rb
@@ -7,23 +7,17 @@ require 'active_support/all'
 require_relative 'instructor/version'
 require_relative 'instructor/openai/patch'
 require_relative 'instructor/openai/response'
+require_relative 'instructor/openai/mode'
 require_relative 'instructor/anthropic/patch'
 require_relative 'instructor/anthropic/response'
-require_relative 'instructor/mode'
 
 # Instructor makes it easy to reliably get structured data like JSON from Large Language Models (LLMs)
 # like GPT-3.5, GPT-4, GPT-4-Vision
 module Instructor
-  @mode = nil
-
   class Error < ::StandardError; end
 
   # The ValidationError class represents an error that occurs during validation.
   class ValidationError < ::StandardError; end
-
-  def self.mode
-    @mode
-  end
 
   # Patches the OpenAI client to add the following functionality:
   # - Retries on exceptions
@@ -34,7 +28,7 @@ module Instructor
   # @param mode [Symbol] The mode to be used. Default is `Instructor::Mode::TOOLS.function`.
   # @return [OpenAI::Client] The patched OpenAI client.
   def self.from_openai(openai_client, mode: :structured_output)
-    @mode = mode
+    Instructor::OpenAI.mode = mode
     openai_client.prepend(Instructor::OpenAI::Patch)
   end
