+++ spec/instructor_spec.rb
@@ -5,11 +5,16 @@ require 'spec_helper'
 RSpec.describe Instructor, '.class' do
   it 'returns the default mode after patching' do
     described_class.from_openai(OpenAI::Client)
-    expect(described_class.mode).to eq(Instructor::Mode::TOOLS.function)
+    expect(Instructor::OpenAI.mode).to eq(:structured_output)
   end
 
-  it 'changes the the mode' do
-    described_class.from_openai(OpenAI::Client, mode: Instructor::Mode::TOOLS.auto)
-    expect(described_class.mode).to eq(Instructor::Mode::TOOLS.auto)
+  it 'changes the the mode to structured output' do
+    described_class.from_openai(OpenAI::Client, mode: :structured_output)
+    expect(Instructor::OpenAI.mode).to eq(:structured_output)
+  end
+
+  it 'changes the the mode to function calling' do
+    described_class.from_openai(OpenAI::Client, mode: :function_calling)
+    expect(Instructor::OpenAI.mode).to eq(:function_calling)
   end
 end
